refactor(client): extract pure-render helpers from chat_panel.dart
